Removed block.hpp and refactored
