+++ include/xsi_factory.hpp
@@ -27,7 +27,6 @@
 
 #include <stdint.h>
 
-#include "block.hpp"
 #include "fs.hpp"
 
 #include "gt_block.hpp"
